etcdserver: add new raft internal request DowngradeInfoSetRequest
