+++ etcdserver/api/membership/membershippb/membership.proto
@@ -36,3 +36,8 @@ message ClusterMemberAttrSetRequest {
   uint64 member_ID = 1;
   Attributes member_attributes = 2;
 }
+
+message DowngradeInfoSetRequest {
+  bool enabled = 1;
+  string ver = 2;
+}
