+++ etcdserver/api/membership/store.go
@@ -90,43 +90,6 @@ func mustSaveDowngradeToBackend(lg *zap.Logger, be backend.Backend, downgrade *D
 	tx.UnsafePut(clusterBucketName, dkey, dvalue)
 }
 
-func downgradeInfoFromBackend(lg *zap.Logger, be backend.Backend) *DowngradeInfo {
-	dkey := backendDowngradeKey()
-	if be != nil {
-		tx := be.ReadTx()
-		tx.Lock()
-		defer tx.Unlock()
-		_, vs := tx.UnsafeRange(clusterBucketName, dkey, nil, 0)
-
-		if len(vs) != 0 {
-			var d DowngradeInfo
-			if err := json.Unmarshal(vs[0], &d); err != nil {
-				if lg != nil {
-					lg.Panic("failed to unmarshal downgrade information", zap.Error(err))
-				}
-			}
-			return &d
-		}
-	}
-	return nil
-}
-
-func clusterVersionFromBackend(be backend.Backend) *semver.Version {
-	ckey := backendClusterVersionKey()
-	if be != nil {
-		tx := be.ReadTx()
-		tx.Lock()
-		defer tx.Unlock()
-
-		_, vs := tx.UnsafeRange(clusterBucketName, ckey, nil, 0)
-		if len(vs) != 0 {
-			v := string(vs[0])
-			return semver.Must(semver.NewVersion(v))
-		}
-	}
-	return nil
-}
-
 func mustSaveMemberToStore(s v2store.Store, m *Member) {
 	b, err := json.Marshal(m.RaftAttributes)
 	if err != nil {
@@ -173,6 +136,123 @@ func mustSaveClusterVersionToStore(s v2store.Store, ver *semver.Version) {
 	if _, err := s.Set(StoreClusterVersionKey(), false, ver.String(), v2store.TTLOptionSet{ExpireTime: v2store.Permanent}); err != nil {
 		plog.Panicf("save cluster version should never fail: %v", err)
 	}
+}
+
+func downgradeInfoFromBackend(lg *zap.Logger, be backend.Backend) *DowngradeInfo {
+	dkey := backendDowngradeKey()
+	if be != nil {
+		tx := be.ReadTx()
+		tx.Lock()
+		defer tx.Unlock()
+		keys, vals := tx.UnsafeRange(clusterBucketName, dkey, nil, 0)
+		if len(keys) == 0 {
+			return nil
+		}
+
+		if len(keys) != 1 {
+			if lg != nil {
+				lg.Panic(
+					"unexpected number of keys when getting cluster version from backend",
+					zap.Int("number-of-key", len(keys)),
+				)
+			}
+		}
+		var d DowngradeInfo
+		if err := json.Unmarshal(vals[0], &d); err != nil {
+			if lg != nil {
+				lg.Panic("failed to unmarshal downgrade information", zap.Error(err))
+			}
+		}
+		return &d
+	}
+	return nil
+}
+
+func clusterVersionFromBackend(lg *zap.Logger, be backend.Backend) *semver.Version {
+	ckey := backendClusterVersionKey()
+	tx := be.ReadTx()
+	tx.RLock()
+	defer tx.RUnlock()
+	keys, vals := tx.UnsafeRange(clusterBucketName, ckey, nil, 0)
+	if len(keys) == 0 {
+		return nil
+	}
+	if len(keys) != 1 {
+		if lg != nil {
+			lg.Panic(
+				"unexpected number of keys when getting cluster version from backend",
+				zap.Int("number-of-key", len(keys)),
+			)
+		}
+	}
+	return semver.Must(semver.NewVersion(string(vals[0])))
+}
+
+func membersFromStore(lg *zap.Logger, st v2store.Store) (map[types.ID]*Member, map[types.ID]bool) {
+	members := make(map[types.ID]*Member)
+	removed := make(map[types.ID]bool)
+	e, err := st.Get(StoreMembersPrefix, true, true)
+	if err != nil {
+		if isKeyNotFound(err) {
+			return members, removed
+		}
+		if lg != nil {
+			lg.Panic("failed to get members from store", zap.String("path", StoreMembersPrefix), zap.Error(err))
+		} else {
+			plog.Panicf("get storeMembers should never fail: %v", err)
+		}
+	}
+	for _, n := range e.Node.Nodes {
+		var m *Member
+		m, err = nodeToMember(n)
+		if err != nil {
+			if lg != nil {
+				lg.Panic("failed to nodeToMember", zap.Error(err))
+			} else {
+				plog.Panicf("nodeToMember should never fail: %v", err)
+			}
+		}
+		members[m.ID] = m
+	}
+
+	e, err = st.Get(storeRemovedMembersPrefix, true, true)
+	if err != nil {
+		if isKeyNotFound(err) {
+			return members, removed
+		}
+		if lg != nil {
+			lg.Panic(
+				"failed to get removed members from store",
+				zap.String("path", storeRemovedMembersPrefix),
+				zap.Error(err),
+			)
+		} else {
+			plog.Panicf("get storeRemovedMembers should never fail: %v", err)
+		}
+	}
+	for _, n := range e.Node.Nodes {
+		removed[MustParseMemberIDFromKey(n.Key)] = true
+	}
+	return members, removed
+}
+
+func clusterVersionFromStore(lg *zap.Logger, st v2store.Store) *semver.Version {
+	e, err := st.Get(path.Join(storePrefix, "version"), false, false)
+	if err != nil {
+		if isKeyNotFound(err) {
+			return nil
+		}
+		if lg != nil {
+			lg.Panic(
+				"failed to get cluster version from store",
+				zap.String("path", path.Join(storePrefix, "version")),
+				zap.Error(err),
+			)
+		} else {
+			plog.Panicf("unexpected error (%v) when getting cluster version from store", err)
+		}
+	}
+	return semver.Must(semver.NewVersion(*e.Node.Value))
 }
 
 // nodeToMember builds member from a key value node.
