+++ tests/e2e/ctl_v3_test.go
@@ -88,22 +88,7 @@ func clusterVersionTest(cx ctlCtx, expected string) {
 	var err error
 	for i := 0; i < 7; i++ {
 		if err = cURLGet(cx.epc, cURLReq{endpoint: "/version", expected: expected}); err != nil {
-			cx.t.Logf("#%d: v3 is not ready yet (%v)", i, err)
-			time.Sleep(time.Second)
-			continue
-		}
-		break
-	}
-	if err != nil {
-		cx.t.Fatalf("failed cluster version test expected %v got (%v)", expected, err)
-	}
-}
-
-func clusterVersionTest(cx ctlCtx, expected string) {
-	var err error
-	for i := 0; i < 7; i++ {
-		if err = cURLGet(cx.epc, cURLReq{endpoint: "/version", expected: expected}); err != nil {
-			cx.t.Logf("#%d:  is not ready yet (%v)", i, err)
+			cx.t.Logf("#%d: version is not ready yet (%v)", i, err)
 			time.Sleep(time.Second)
 			continue
 		}
